requirements.txt
asgiref==3.2.10
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-autoslug==1.9.8
django-cleanup==6.0.0
django-environ==0.8.1
django-mathfilters==1.0.0
django-rosetta==0.9.8
gunicorn==20.1.0
h11==0.13.0
idna==3.3
nose==1.3.7
outcome==1.1.0
Pillow==9.0.1
polib==1.1.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
Random-Word==1.0.7
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
Unidecode==1.3.4
urllib3==1.26.9
whitenoise==6.0.0
wsproto==1.1.0
